Refactor evm events and ExamplePlugins
Changes:
* Removed `_evm_` from EVM's event names
* `did/will_execute_instruction` is now merged with
`did/will_evm_execute_instruction` (the old - native-like event is not
needed in evm)
* Created `ExamplePlugin` for both EVM and Native
